Fix Charles comments
