+++ pyproject.toml
@@ -206,7 +206,7 @@ scil_sh_fusion  = "scilpy.cli.scil_sh_fusion:main"
 scil_sh_to_aodf  = "scilpy.cli.scil_sh_to_aodf:main"
 scil_sh_to_rish  = "scilpy.cli.scil_sh_to_rish:main"
 scil_sh_to_sf  = "scilpy.cli.scil_sh_to_sf:main"
-scil_fodf_global_sf_threshold = "scilpy.cli.scil_fodf_global_sf_threshold:main"
+scil_fodf_threshold_by_amplitude = "scilpy.cli.scil_fodf_threshold_by_amplitude:main"
 scil_stats_group_comparison  = "scilpy.cli.scil_stats_group_comparison:main"
 scil_surface_assign_custom_color  = "scilpy.cli.scil_surface_assign_custom_color:main"
 scil_surface_assign_uniform_color  = "scilpy.cli.scil_surface_assign_uniform_color:main"
