+++ src/scilpy/cli/scil_tracking_local_dev.py
@@ -82,7 +82,8 @@ from scilpy.tracking.seed import CustomSeedsDispenser, SeedGenerator
 from scilpy.tracking.tracker import Tracker
 from scilpy.tracking.utils import (add_mandatory_options_tracking,
                                    add_out_options, add_seeding_options,
-                                   add_tracking_options, get_theta,
+                                   add_tracking_options,
+                                   get_global_sf_threshold_mask, get_theta,
                                    verify_seed_options,
                                    verify_streamline_length_options)
 from scilpy.version import version_string
@@ -308,7 +309,6 @@ def main():
 
         if args.global_sf_rel_thr is not None or \
                 args.global_sf_abs_thr is not None:
-            from scilpy.tracking.utils import get_global_sf_threshold_mask
             sf_mask = get_global_sf_threshold_mask(
                 odf_sh_data, args, sh_basis, is_legacy)
