+++ src/scilpy/cli/tests/test_tracking_local.py
@@ -202,3 +202,23 @@ def test_execution_tracking_fodf_custom_seeds(script_runner, monkeypatch):
                              '--compress', '0.1', '--sh_basis', 'descoteaux07',
                              '--min_length', '20', '--max_length', '200'])
     assert ret.success
+
+
+def test_execution_tracking_global_sf_threshold(script_runner, monkeypatch):
+    monkeypatch.chdir(os.path.expanduser(tmp_dir.name))
+    in_fodf = os.path.join(SCILPY_HOME, 'tracking', 'fodf.nii.gz')
+    in_mask = os.path.join(SCILPY_HOME, 'tracking', 'seeding_mask.nii.gz')
+
+    # Test with relative threshold
+    ret = script_runner.run(['scil_tracking_local', in_fodf,
+                             in_mask, in_mask, 'global_sf_rel.trk',
+                             '--nt', '10', '--global_sf_rel_thr', '0.1',
+                             '--sh_basis', 'descoteaux07'])
+    assert ret.success
+
+    # Test with absolute threshold
+    ret = script_runner.run(['scil_tracking_local', in_fodf,
+                             in_mask, in_mask, 'global_sf_abs.trk',
+                             '--nt', '10', '--global_sf_abs_thr', '0.05',
+                             '--sh_basis', 'descoteaux07', '-f'])
+    assert ret.success
